+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 - Relational fields now load elements in the current site rather than the primary site, if the source element isn’t localizable. ([#7048](https://github.com/craftcms/cms/issues/7048))
 - `craft\base\Model::setAttributes()` now normalizes date attributes into `DateTime` objects.
 - `craft\services\Gql::getValidationRules()` now has an `$isIntrospectionQuery` argument.
-- Updated Twig to 2.13.
+- Updated Twig to 3.1.
 - Updated LitEmoji ot 2.x.
 - Updated `webonyx/graphql-php` to 14.1.1.